server/requirements.txt
# Dependencies for Video Analytics API Server
# Install with: pip install -r requirements.txt

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9

# Validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0
